+++ Modules/Intent.Modules.Entities.Repositories.Api/release-notes.md
@@ -1,3 +1,7 @@
+### Version 5.1.3
+
+- Fixed: template config issue for `EntityRepositoryInterface`.
+
 ### Version 5.1.2
 
 - Improvement: Updated module NuGet packages infrastructure.